Updated index.coffee with changes for brunch 1.7
I've changed the json2css probles and changed the config options reading as they were taken from the old format of configuration file of brunch. 

I've added a dummy data for the mixins for json2css not ideal but if anyting else is suggested let me know i can change it.